Trait com_rs::Unknown

pub trait Unknown {
    unsafe fn query_interface(&self, iid: REFIID, object: *mut LPVOID) -> HRESULT { ... }
    unsafe fn add_ref(&self) -> ULONG { ... }
    unsafe fn release(&self) -> ULONG { ... }
}

Base trait for all COM types.

Provided Methods

unsafe fn query_interface(&self, iid: REFIID, object: *mut LPVOID) -> HRESULT

Retrieves pointers to the supported interfaces on an object. Use ComPtr::from instead.

unsafe fn add_ref(&self) -> ULONG

Increments the reference count for an interface on an object. Should never need to call this directly.

unsafe fn release(&self) -> ULONG

Decrements the reference count for an interface on an object. Should never need to call this directly.